Moved hatcehet to top level external directory.
