+++ hatchet/external/__init__.py
@@ -2,3 +2,7 @@
 # Hatchet Project Developers. See the top-level LICENSE file for details.
 #
 # SPDX-License-Identifier: MIT
+
+from .roundtrip.roundtrip.manager import Roundtrip
+
+Roundtrip
